feat!: altera estrutura de pastas do código
